Hammerhead: defconfig: First Build 3.10.40 "stable"
Why say "stable"? So...

The kernel at this point, work's very well. I add what work's and what don't work's

- Display
- Backlight
- Graphics
- Video-Playback
- USB storage
- Bluetooth
- Audio
- Vibrator
- Battery
- Charger
- Sim "only Call and SMS"

This are the most thing's that work's very well.

What's won't work...

- USB adb for debugging (MOST IMPORTANT TO FIX)
- Rotator
- Wifi
- 3G-4G

And other thing's that at this moment, I don't know because debug is broke and
I don't see what is also broke.

This I think is the best base to use for bring-up 3.10 on hammerhead device.

Signed-off-by: Emanuele Scarlata (Fenix46) <scarlataemanuele@gmail.com>